refactor: split main.py into router modules
Extracts all domain routes from the 832-line main.py monolith into:
- routers/factors.py  — factor CRUD + theme/version read endpoints
- routers/claims.py   — claim CRUD + theme/version read endpoints
- routers/organizations.py — org/user/invite management
- routers/hierarchy.py — project/theme/version/space hierarchy

main.py now contains only app init, middleware, lifespan,
/health, /chat and WebSocket. All 47 auth coverage tests pass.

Closes #283
